+++ lib/less/rails.rb
@@ -3,6 +3,11 @@ module Less
   end
 end
 
+require 'rails'
+require 'tilt'
+require 'sprockets'
+require 'sprockets/railtie'
+
 require 'less/rails/version'
 require 'less/rails/template_handlers'
 require 'less/rails/railtie'