+++ lib/less/rails/railtie.rb
@@ -9,6 +9,7 @@ module Less
       config.less = ActiveSupport::OrderedOptions.new
       
       config.less.paths = []
+      config.less.options = {}
       
       config.before_initialize do |app|
         unless app.config.assets && app.config.assets.enabled