+++ test/cases/basics_spec.rb
@@ -1,0 +1,11 @@
+require 'test_helper'
+
+class BasicsSpec < Less::Rails::Spec
+  
+  let(:basics) { dummy_asset('basics') }
+  
+  it 'must render variables' do
+    basics.must_match %r{#test-variable\{color:#4d926f;\}}
+  end
+
+end