+++ test/cases/railtie_spec.rb
@@ -1,0 +1,41 @@
+require 'test_helper'
+
+class RailtieSpec < Less::Rails::Spec
+  
+  describe 'config' do
+    
+    it 'must have a less ordered hash' do
+      dummy_config.less.must_be_instance_of ActiveSupport::OrderedOptions
+    end
+    
+    it 'must have an empty array for paths' do
+      dummy_config.less.paths.must_equal Array.new
+    end
+    
+    it 'must have an options hash passed down to the #to_css method' do
+      dummy_config.less.options.must_be_instance_of Hash
+      basic_compressed_match = /#test-variable\{color:#4d926f;\}/
+      dummy_config.less.options[:compress] = true
+      dummy_asset('basics').must_match basic_compressed_match
+      reset_assets_cache
+      dummy_config.less.options[:compress] = false
+      dummy_asset('basics').wont_match basic_compressed_match
+    end
+  
+  end
+  
+  describe 'initialization' do
+    
+    it 'must register our template engine' do
+      dummy_assets.engines['.less'].must_equal Less::Rails::LessTemplate
+    end
+    
+    it 'must extend the context class with our config' do
+      dummy_assets.context_class.must_respond_to :less_config
+      dummy_assets.context_class.less_config.must_equal dummy_config.less
+    end
+    
+  end
+
+
+end